move lockfile helper types together with util rules
resolves runtime import
